refactor(components): merge useFavoriteLocation \& useFillStar
МЕТА ДАННЫЕ: refactor 13
