+++ src/hooks/useFavoriteLocation.js
@@ -1,7 +1,7 @@
-import { computed, ref } from 'vue';
+import { computed, onMounted, ref } from 'vue';
 import { useStore } from 'vuex';
 
-export default function useFavoriteLocation(fillStar) {
+export default function useFavoriteLocation() {
   const store = useStore();
   const favoriteLocations = ref({});
   const favoriteLocationsLength = ref(0);
@@ -104,6 +104,31 @@ export default function useFavoriteLocation(fillStar) {
       store.commit('favoriteCity/setShowModal', false);
     }, 1500);
   };
+
+  const fillStar = () => {
+    favoriteLocations.value = store.state.favoriteCity.favoriteLocations;
+    chosenLocation.value = store.state.selectedCity.chosenCity.name;
+    favoriteLocationsLength.value = Object.keys(
+      favoriteLocations.value
+    ).length;
+    console.log('length', favoriteLocationsLength.value);
+
+    if (favoriteLocationsLength.value) {
+      let obj = [];
+      for (let i = 0; i < favoriteLocationsLength.value; i++) {
+        obj = Object.values(favoriteLocations.value)[i];
+        console.log('obj', obj?.name);
+        if (obj?.name === chosenLocation.value) {
+          store.commit('favoriteCity/setCityFavorite', true);
+          return;
+        } else {
+          store.commit('favoriteCity/setCityFavorite', false);
+        }
+      }
+    }
+  };
+
+  onMounted(fillStar);
 
   return {
     favoriteLocation,
